Improved code structure.
