+++ Assets/Prefabs/Enemies/SimpleEnemy.prefab
@@ -1,7 +1,7 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!65 &2106045691564816451
-BoxCollider:
+--- !u!135 &1200612364235314869
+SphereCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
@@ -11,7 +11,7 @@ BoxCollider:
   m_IsTrigger: 0
   m_Enabled: 1
   serializedVersion: 2
-  m_Size: {x: 0.12354162, y: 0.049493548, z: 0.050257318}
+  m_Radius: 0.025
   m_Center: {x: 0.0000000037252903, y: 0.00043681628, z: 0.003158066}
 --- !u!1 &7217388690386080259
 GameObject:
@@ -61,6 +61,11 @@ MonoBehaviour:
   maxMovSpeed: 30
   acceleration: 30
   angularSpeed: 30
+  minTargetDistance: 10
+  maxTargetDistance: 30
+  maxStrafeDistanceBias: 5
+  timeBetweenShots: 2
+  behaviorType: 0
 --- !u!54 &6922267231102003771
 Rigidbody:
   m_ObjectHideFlags: 0
