+++ Assets/Scripts/Enemies/Behaviors/ExplodingBehavior.cs
@@ -1,5 +1,7 @@
 using System;
 using UnityEngine;
+using Enemies.Behaviors;
+using Enemies.Tasks;
 
 namespace Enemies
 {
