+++ Assets/Scripts/Enemies/Tasks/ExplodeTask.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Enemies
+namespace Enemies.Tasks
 {
     /// <summary>
     /// Task representing exploding to damage specified target.
