+++ Assets/Scripts/Enemies/Tasks/MoveToTask.cs
@@ -1,24 +1,23 @@
 using UnityEngine;
 
-namespace Enemies
+namespace Enemies.Tasks
 {
-    public class SpinMoveToTask : MoveToTask
+    /// <summary>
+    /// Task representing continuous movement toward a certain target.
+    /// </summary>
+    public class MoveToTask : Task
     {
-        public EnemyController.RotationDirection SpinDirection = EnemyController.RotationDirection.Clockwise;
-        
+        public float DistanceThreshold { get; set; } = 1f;
+
         public override void ExecuteUpdate(EnemyController controller)
         {
             Vector3 currentPosition = controller.transform.position;
             Vector3 targetPosition = currentPosition;
-
-            if (Target != null)
-            {
-                targetPosition = Target.GetPosition();
-            }
+            if (Target != null) { targetPosition = Target.GetPosition(); }
 
             if (Vector3.Distance(currentPosition, targetPosition) > DistanceThreshold)
             {
-                controller.UpdateSpinMoveTo(targetPosition, SpinDirection);
+                controller.UpdateMoveTo(targetPosition);
             }
             else
             {
