+++ Assets/Scripts/Enemies/Tasks/SpinStrafeTask.cs
@@ -1,27 +1,23 @@
 using UnityEngine;
 
-namespace Enemies
+namespace Enemies.Tasks
 {
     /// <summary>
-    /// Task representing continuous rotating to face a certain target.
+    /// Task representing movement around the specified Target while spinning.
     /// </summary>
-    public class KeepRotatedTowardsTask : Task
+    public class SpinStrafeTask : StrafeTask
     {
-        public float DistanceThreshold { get; set; } = 1f;
-        
+        public EnemyController.RotationDirection SpinDirection = EnemyController.RotationDirection.Clockwise;
+
         public override void ExecuteUpdate(EnemyController controller)
         {
             Vector3 currentPosition = controller.transform.position;
             Vector3 targetPosition = currentPosition;
-
-            if (Target != null)
-            {
-                targetPosition = Target.GetPosition();
-            }
+            if (Target != null) { targetPosition = Target.GetPosition(); }
             
             if (Vector3.Distance(currentPosition, targetPosition) <= DistanceThreshold)
             {
-                controller.UpdateKeepRotatedTowards(targetPosition);
+                controller.UpdateSpinStrafe(targetPosition, StrafeDirection, SpinDirection);
             }
             else
             {
