+++ Assets/Scripts/Enemies/Tasks/StrafeTask.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Enemies
+namespace Enemies.Tasks
 {
     /// <summary>
     /// Task representing movement around the specified Target.
@@ -15,11 +15,7 @@ namespace Enemies
         {
             Vector3 currentPosition = controller.transform.position;
             Vector3 targetPosition = currentPosition;
-
-            if (Target != null)
-            {
-                targetPosition = Target.GetPosition();
-            }
+            if (Target != null) { targetPosition = Target.GetPosition(); }
             
             if (Vector3.Distance(currentPosition, targetPosition) <= DistanceThreshold)
             {
