+++ Assets/Scripts/Enemies/Tasks/WaitTask.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Enemies
+namespace Enemies.Tasks
 {
     /// <summary>
     /// Task representing waiting.
